AI_assistant/requirements.txt
# Core Flask and web framework
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.11.0
gunicorn == 23.0.0


# embedding
chromadb==1.0.9

# LangChain and AI dependencies
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59t
langchain-openai==0.3.16
langchain-text-splitters==0.3.8

# OpenAI
openai==1.78.1

# HTTP requests (if not already included by other packages)
requests==2.32.3

# Configuration management (if you use python-dotenv)
python-dotenv==1.0.1

# JWT authentication for Auth0
PyJWT==2.8.0
cryptography==41.0.7